+++ jungrapht-visualization-samples/src/main/java/org/jungrapht/samples/LensDemo.java
@@ -121,7 +121,7 @@ public class LensDemo extends JPanel {
     vv.getRenderContext().setVertexLabelFunction(Object::toString);
 
     final Function<String, Shape> ovals = vv.getRenderContext().getVertexShapeFunction();
-    final Function<String, Shape> squares = n -> new Rectangle2D.Float(-10, -10, 20, 20);
+    final Function<String, Shape> squares = n -> new Rectangle2D.Double(-10, -10, 20, 20);
 
     // add a listener for ToolTips
     vv.setVertexToolTipFunction(n -> n); //Object::toString);
